Add files via upload
changed various things- worked further with menu bar and links on page- changed top link to go to site instead of company GitHub. 

Problems that are still persistent- when window is narrow, the navbar goes over the site logo, blocking the link there. Mobile usage of the site is likely compromised by this.  the logo at the top also re-lines itself and goes down onto the picture of Mattoon.  this is almost as ugly as the entire rest of the page combined. 

when window is wide, Mattoon's face is stretched out significantly- we also need a better image of him, refer to the project section, ill see what i can do there.